Remove uncessary and incorrect copyright statement from the footer
The copyright statement in the footer of the site had the incorrect date and in
the context of Bulgarian law and the specific purpose of the site is at best
misleading if not completely wrong.

This commit simplifies it and leaves only the internationalized part of it which
currently contains only the title of the site.